pyproject.toml
[project]
name = "knowledge-base"
version = "0.1.0"
description = "AI-powered knowledge base with semantic search and RAG capabilities"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "redis>=5.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "duckdb>=1.0.0",
    "tenacity>=8.2.0",
    "click>=8.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.26.0",
    "aiohttp>=3.9.0",
    "starlette>=0.27.0",
    "markdownify>=0.11.0",
    "google-cloud-aiplatform>=1.38.0",
    "rank-bm25>=0.2.2",
    "pypdf>=3.17.0",
    "anthropic[vertex]>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "httpx>=0.26.0",
]

[project.scripts]
kb = "knowledge_base.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "e2e: marks tests as end-to-end tests (require live Slack)",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
